requirements.txt
discord.py==2.3.2
streamlit==1.31.0
python-dotenv==1.0.0